Move CLI layout constants into domain
Keep the CLI and skill layout contract with the domain layer so application services can stay focused on orchestration instead of owning tool-specific naming and distribution rules.
