+++ Packages/src/Editor/Infrastructure/CLI/CliContractReader.cs
@@ -2,7 +2,7 @@ using System.IO;
 
 using Newtonsoft.Json.Linq;
 
-using io.github.hatayama.UnityCliLoop.Application;
+using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
 
 namespace io.github.hatayama.UnityCliLoop.Infrastructure
