+++ Packages/src/Editor/Infrastructure/CLI/CliInstallationDetector.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using UnityEngine;
 
 using io.github.hatayama.UnityCliLoop.Application;
+using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
 
 namespace io.github.hatayama.UnityCliLoop.Infrastructure
