+++ Packages/src/Editor/Infrastructure/CLI/ProjectLocalCliAutoInstaller.cs
@@ -2,6 +2,7 @@ using UnityEditor;
 using UnityEngine;
 
 using io.github.hatayama.UnityCliLoop.Application;
+using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
 
 namespace io.github.hatayama.UnityCliLoop.Infrastructure
