+++ Packages/src/Editor/Infrastructure/CLI/ProjectLocalCliInstaller.cs
@@ -6,6 +6,7 @@ using System.Security.Cryptography;
 using UnityEngine;
 
 using io.github.hatayama.UnityCliLoop.Application;
+using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
 
 namespace io.github.hatayama.UnityCliLoop.Infrastructure
