+++ Packages/src/Editor/Infrastructure/SkillSetup/SkillInstallLayout.cs
@@ -6,7 +6,6 @@ using System.Text.RegularExpressions;
 
 using UnityEngine;
 
-using io.github.hatayama.UnityCliLoop.Application;
 using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
